fix(native-player): avoid boxed/overlay subtitle overlap on PiP exit
On leaving PiP the overlay was revealed in the same turn that re-enabled
suppression, so the still-visible boxed caption and the overlay briefly
showed together. Keep the overlay hidden until the native caption has been
flushed (deselect/reselect), then reveal it.

Also move the self-contained subtitle rendering — SubtitleStyle, SubtitleRun
and SubtitleOverlayView, plus the hex-colour and cue-flattening helpers — out
of NativePlayerPlugin into SubtitleOverlayView.swift. No behaviour change from
the move; the plugin keeps only the AVPlayer plumbing.
